backend/requirements.txt
django~=4.2
django-oscar[sorl-thumbnail]==3.2.4
django-oscar-api==3.2.4
pillow==10.2.0
drf-spectacular==0.27.1
aiogram==3.3.0
python-dotenv==1.0.1
psycopg==3.1.18
psycopg[binary,c]==3.1.18
django-cors-headers==4.3.1
gunicorn==21.2.0
djangorestframework-simplejwt==5.3.1
sentry-sdk[django]==1.40.2
django-filter==23.5
pydantic-settings==2.1.0
celery==5.3.6
redis==5.0.1
selenium==4.18.1
webdriver-manager==4.0.1
email-validator==2.1.1
yookassa==3.0.1
pandas==2.2.0
openpyxl==3.1.2
numpy==1.26.3
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
et-xmlfile==1.1.0